Discard own messages
Unifying all attack messages in a sigle topics obey the system to
filtering the desired messages and avoid it own messages.

A small fix were made on message being produced once the value was being
sent rather than message which provides the key - a crucial config to
properly rotate the message to a specific partition.